ENH,CLN: add login context provider
