+++ frontend/src/context/DefaultState.tsx
@@ -1,52 +1,4 @@
-import * as React from 'react';
-import { ISlackContextProvider, ISiweProps } from './types.d';
-
-/*
-Context list
-1. slack config
-2. wallet key pair
-3. owlly login info
-4. dropdown state
-*/
-export const SlackContext = React.createContext({
-  slackContext: {
-    triggerName: '',
-    botOauthToken: '',
-    channelID: '',
-    userID: '',
-    userName: '',
-  },
-
-  // type Dispatch<A> = (value: A) => void;
-  setSlackContext: (
-    prev: React.SetStateAction<{
-      triggerName: string;
-      botOauthToken: string;
-      channelID: string;
-      userID: string;
-      userName: string;
-    }>
-  ) => {},
-});
-
-export function SlackContextProvider({ children }: ISlackContextProvider) {
-  const [slackContext, setSlackContext] = React.useState({
-    triggerName: '',
-    botOauthToken: '',
-    channelID: '',
-    userID: '',
-    userName: '',
-  });
-
-  const value = { slackContext, setSlackContext };
-
-  // prettier-ignore
-  return (
-    <SlackContext.Provider value={value}>
-      {children}
-    </SlackContext.Provider>
-  );
-}
+import { ISiweProps } from './types.d';
 
 export const breakpoints = {
   device: {
